Set up User interface and transfer tokens

+++ src/contracts/DaiTokenMock.sol
@@ -6,7 +6,7 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract DaiTokenMock is ERC20{
   
 
-  constructor(uint initialSupply) public ERC20("Dai StableCoin", "DAI") {
+  constructor(uint256 initialSupply) public ERC20("Dai StableCoin", "DAI") {
     _mint(msg.sender, initialSupply);
   }
 }
